Don't render an empty list of Contributors
